+++ glimpse/core/src/commonMain/kotlin/lenses/PerspectiveLens.kt
@@ -28,7 +28,7 @@ import graphics.glimpse.types.perspective
  * field of view.
  */
 data class PerspectiveLens(
-    val fovY: Angle,
+    val fovY: Angle<Float>,
     val aspect: Float,
     val near: Float,
     val far: Float