+++ glimpse/core/src/commonMain/kotlin/types/AngleRange.kt
@@ -19,7 +19,7 @@ package graphics.glimpse.types
 /**
  * A range of values of type [Angle].
  */
-data class AngleRange(
-    override val start: Angle,
-    override val endInclusive: Angle
-) : ClosedRange<Angle>
+data class AngleRange<T : Number>(
+    override val start: Angle<T>,
+    override val endInclusive: Angle<T>
+) : ClosedRange<Angle<T>>